Only allow audio tracks for RandomPlay
Embedded cuesheet FLACs (FEC) are inserted into the `tracks` table with
`audio` set to 0. The individual tracks in a FEC are inserted into the
`tracks` table as FLACs with `audio` set to 1. However, FEC files are
not excluded as potential RandomPlay tracks, and will not correctly
populate in the playlist if they are randomly selected. This is
typically seen as the RandomPlay-generated playlist having less than the
expected number of tracks. The issue is mitigated by adding a genre
conditional; FEC files are not associated from genres and therefore are
eliminated as potential RandomPlay tracks.

To address this, I required that RandomPlay tracks must have `audio` set
to 1. I opted to not extend the list of precluded content types for a
couple of reasons:
* I wouldn't expect other, non-excluded playlists from being used in
  RandomPlay (cue, m3u, etc).
* Precedent in Slim::Web::Pages::Search::advancedSearch to append
  `audio = 1`
* Precedent in other modules to limit the precluded content types to
  `cpl`, `src`, `ssp`, and `dir`