Persist update URL in prefs. This allows helpers to still get the update, even if auto download is disabled in the server.
Reset status on start/stop to only flag an update if LMS is running and has checked for an update.

Add some code to help my local testing when running from source. This usually would disable the update check, as there's no update to a repository clone. This little code allows me to still check for updates on macOS with a simple flag.